+++ examples/actor.rs
@@ -1,3 +1,4 @@
+use std::error::Error;
 use std::process::exit;
 use std::time::Duration;
 use tyra::prelude::{Actor, ActorContext, ActorFactory, ActorMessage, ActorResult, ActorSystem, Handler, TyraConfig};
@@ -31,11 +32,11 @@ struct HelloWorldFactory {
 }
 
 impl ActorFactory<HelloWorld> for HelloWorldFactory {
-    fn new_actor(&self, _context: ActorContext<HelloWorld>) -> HelloWorld {
-        HelloWorld {
+    fn new_actor(&mut self, _context: ActorContext<HelloWorld>) -> Result<HelloWorld, Box<dyn Error>> {
+        Ok(HelloWorld {
             count: self.count,
             text: self.text.clone(),
-        }
+        })
     }
 }
 impl Handler<MessageA> for HelloWorld {