+++ examples/delayed.rs
@@ -1,3 +1,4 @@
+use std::error::Error;
 use std::process::exit;
 use std::thread::sleep;
 use std::time::Duration;
@@ -45,8 +46,8 @@ impl TemplateActorFactory {
 }
 
 impl ActorFactory<TemplateActor> for TemplateActorFactory {
-    fn new_actor(&self, _context: ActorContext<TemplateActor>) -> TemplateActor {
-        TemplateActor::new()
+    fn new_actor(&mut self, _context: ActorContext<TemplateActor>) -> Result<TemplateActor, Box<dyn Error>> {
+        Ok(TemplateActor::new())
     }
 }