+++ src/system/delay_actor.rs
@@ -1,3 +1,4 @@
+use std::error::Error;
 use std::thread::sleep;
 use std::time::Duration;
 use log::debug;
@@ -14,8 +15,8 @@ impl DelayActor {
 
 pub struct DelayActorFactory {}
 impl ActorFactory<DelayActor> for DelayActorFactory {
-    fn new_actor(&self, _context: ActorContext<DelayActor>) -> DelayActor {
-        DelayActor::new()
+    fn new_actor(&mut self, _context: ActorContext<DelayActor>) -> Result<DelayActor, Box<dyn Error>> {
+        return Ok(DelayActor::new());
     }
 }
 impl DelayActorFactory {